April 30 - UCI/Perft Development
Engine development for April 30, 2025.

Created functions for parsing move and position commands sent from UCI to engine, and created skeleton for a basic perft function that counts the amount of nodes visited.

Plan for next iteration of development is to : 
- Finish functions for parsing UCI commands and enable engine to communicate with UCI
- Use Lichess-bot to enable my function to function as a Lichess bot which can be played against.
-Add more detail to perft debugging functions
After these implementations, next on the list is to make major improvements to the search and evaluation functions, particularly evaluation as it is currently a skeleton if anything

Afterwards, code cleanup/refactoring is next. I plan to add more comments throughout the code and potentially rewrite code.

Potentially, I may split my code into different c/header files to improve the refactoring process.